__MAIN__/database_scripts/create_database.py
import sys
import os
sys.path.insert(0, os.path.dirname(__file__).replace('\\', '/') + '/../')

from mysql.connector import MySQLConnection
from mysql.connector.abstracts import MySQLConnectionAbstract, MySQLCursorAbstract
from config import database_config

def create_database() -> None:
    connection: MySQLConnectionAbstract
    with MySQLConnection(**database_config) as connection:
        connection.autocommit = True
        cursor: MySQLCursorAbstract
        with connection.cursor(dictionary=True) as cursor:
            cursor.execute(f"DROP DATABASE IF EXISTS {database_config['database']}")
            cursor.execute(f"CREATE DATABASE {database_config['database']}")

if __name__ == '__main__':
    create_database()